deploy/requirements.txt
torch>=1.12.0
evdev
sshkeyboard
pyyaml>=6.0
streamlit>=1.28.0
numpy>=1.21.0
